pyproject.toml
[tool.poetry]
name = "fiberdiameter"
version = "0.8.1"
description = "Measure the diameter of fibers 10x faster!"
authors = [
            "Fernando Cossio",
            "Yael Suarez",
            "Luis Cossio",
            "Alejandra De Luna"]
license = "Apache License 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
torch = "^1.12.1"
pytorch-lightning = "^1.7.7"
wandb = "^0.13.3"
Pillow = "^9.2.0"
numpy = "^1.23.3"
onnx = "^1.12.0"
pyyaml = "^6.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"
flake8 = "^5.0.4"
black = "^22.8.0"
isort = "^5.10.1"
pytest = "^7.1.3"
pytest-cov = "^3.0.0"
Sphinx = "^5.1.1"
sphinx-autoapi = "^1.9.0"
sphinx-rtd-theme = "^1.0.0"
myst-nb = "^0.16.0"
python-semantic-release = "^7.31.4"
huggingface-hub = "^0.9.1"
jupyter = "^1.0.0"
wandb = "^0.13.3"
lwbna-unet = "^1.0.2"
Pillow = "^9.2.0"
numpy = "^1.23.3"
matplotlib = "^3.6.0"

[tool.semantic_release]
version_variable = "pyproject.toml:version" # version location
branch = "main"                             # branch to make releases of
changelog_file = "CHANGELOG.md"             # changelog file
build_command = "poetry build"              # build dists
dist_path = "dist/"                         # where to put dists
upload_to_release = true                    # auto-create GitHub release
upload_to_pypi = false                      # don't auto-upload to PyPI
remove_dist = false                         # don't remove dists
patch_without_tag = true                    # patch release by default

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
